✅ Add functionality for configuring metrics API URL
